+++ src/app/app.module.ts
@@ -8,19 +8,25 @@ import { CoreModule } from './core/core.module';
 import { CartModule } from './modules/cart/cart.module';
 import { SharedModule } from './shared/shared.module';
 import { ProductsModule } from './modules/products/products.module';
+import { AppRoutingModule } from './app-routing.module';
+import { ContactUsComponent } from './modules/contact-us/contact-us.component';
+import { FormsModule } from '@angular/forms';
 
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    ContactUsComponent
   ],
   imports: [
     BrowserModule,
+    FormsModule,
     CoreModule,
     SharedModule,
     BrowserAnimationsModule,
     CartModule,
-    ProductsModule
+    ProductsModule,
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]