+++ changelog.md
@@ -7,3 +7,6 @@
 - Refactored project structure(shared,core, feature modules)
 - Added functionality for quantity and price calculation
 - Added hover directive for items in cart
+- Added app-routing module
+- Added some demo-services to core module
+- Added demo-services presentation to contact-us page